pyproject.toml
[project]
name = "bank_of_israel_ai"
version = "0.1.0"
description = "A minimal Python project scaffold for Bank of Israel AI initiatives."
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "python-dotenv>=1.0.0",
  "flask>=2.0.0",
  "requests>=2.28.0",
  "google-generativeai>=0.3.0",
  "crewai>=0.11.2",
  "crewai-tools>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
